test/requirements.txt
mysql-connector-python
requests
